Process Qualtrics Data Files

This function reads in and processes a series of CSV files exported from Qualtrics. It assumes the first two rows of each file contain metadata and not actual data.

Usage

process_qualtrics_files(data_files, rename_files = FALSE, new_names = NULL)

Arguments

data_files

A character vector of file paths to the CSV files to be processed.

rename_files

Logical. If TRUE, the data and labels for each file will be saved under a new name based on the original file name. Default is FALSE.

new_names

A character vector of new names for the files. This should be the same length as data_files. If NULL and rename_files is TRUE, new names will be generated by removing the timestamp and anything before it from the original file name. Default is NULL.

Value

This function does not return anything. It writes processed data and labels to CSV files and assigns the resulting data frames to new variables in the global environment.

References

Remember to add reference here

Examples

if (FALSE) {
process_qualtrics_files(c("file1.csv", "file2.csv"), rename_files = TRUE, new_names = c("survey1", "survey2"))
}