Compute Power Analysis For SEM Model Using Monte Carlo Simulation

Compute Power Analysis For SEM Model Using Monte Carlo Simulation

Usage

simulate_power(model.population, model, ksim = 10, nobs = 100)

Arguments

model.population

The lavaan model with population estimate values specified.

model

The lavaan model that will be tested in each simulation

ksim

How many simulations the function should perform

nobs

How many observations to generate in each simulation

Value

A table.

Details

The function uses the simulate_data function from the lavaan package to perform a monte carlo simulation. The mean, std error, z value, p value and confidence intervals are then computed and reported in a table for each parameter.

References

Remember to add reference here

Examples

library(lavaan)
#> This is lavaan 0.6-15
#> lavaan is FREE software! Please report any bugs.
modpop <- '
M ~ 0.40*X
Y ~ 0.30*M
'
mod <- '
M ~ X
Y ~ M
'

simulate_power(modpop, mod)
#>    lhs op rhs   est    se     z pvalue ci.lower ci.upper Parameter
#> 1    M  ~   X 0.310 0.099 3.132  0.002    0.116    0.504     M ~ X
#> 2    Y  ~   M 0.318 0.094 3.382  0.001    0.134    0.502     Y ~ M
#> 3    M ~~   M 0.963 0.136 7.071  0.000    0.696    1.230    M ~~ M
#> 4    Y ~~   Y 0.933 0.132 7.071  0.000    0.675    1.192    Y ~~ Y
#> 5    X ~~   X 0.982 0.000    NA     NA    0.982    0.982    X ~~ X
#> 6    M  ~   X 0.263 0.111 2.384  0.017    0.047    0.480     M ~ X
#> 7    Y  ~   M 0.440 0.095 4.649  0.000    0.254    0.625     Y ~ M
#> 8    M ~~   M 1.032 0.146 7.071  0.000    0.746    1.318    M ~~ M
#> 9    Y ~~   Y 0.975 0.138 7.071  0.000    0.705    1.245    Y ~~ Y
#> 10   X ~~   X 0.845 0.000    NA     NA    0.845    0.845    X ~~ X
#> 11   M  ~   X 0.375 0.096 3.885  0.000    0.186    0.564     M ~ X
#> 12   Y  ~   M 0.301 0.096 3.135  0.002    0.113    0.489     Y ~ M
#> 13   M ~~   M 1.000 0.141 7.071  0.000    0.723    1.277    M ~~ M
#> 14   Y ~~   Y 1.060 0.150 7.071  0.000    0.767    1.354    Y ~~ Y
#> 15   X ~~   X 1.075 0.000    NA     NA    1.075    1.075    X ~~ X
#> 16   M  ~   X 0.543 0.094 5.753  0.000    0.358    0.728     M ~ X
#> 17   Y  ~   M 0.321 0.100 3.226  0.001    0.126    0.516     Y ~ M
#> 18   M ~~   M 0.841 0.119 7.071  0.000    0.608    1.074    M ~~ M
#> 19   Y ~~   Y 1.109 0.157 7.071  0.000    0.802    1.417    Y ~~ Y
#> 20   X ~~   X 0.944 0.000    NA     NA    0.944    0.944    X ~~ X
#> 21   M  ~   X 0.448 0.095 4.693  0.000    0.261    0.635     M ~ X
#> 22   Y  ~   M 0.425 0.095 4.488  0.000    0.240    0.611     Y ~ M
#> 23   M ~~   M 0.859 0.122 7.071  0.000    0.621    1.097    M ~~ M
#> 24   Y ~~   Y 0.941 0.133 7.071  0.000    0.680    1.202    Y ~~ Y
#> 25   X ~~   X 0.944 0.000    NA     NA    0.944    0.944    X ~~ X
#> 26   M  ~   X 0.434 0.086 5.027  0.000    0.265    0.603     M ~ X
#> 27   Y  ~   M 0.343 0.100 3.444  0.001    0.148    0.539     Y ~ M
#> 28   M ~~   M 0.921 0.130 7.071  0.000    0.666    1.177    M ~~ M
#> 29   Y ~~   Y 1.146 0.162 7.071  0.000    0.828    1.464    Y ~~ Y
#> 30   X ~~   X 1.236 0.000    NA     NA    1.236    1.236    X ~~ X
#> 31   M  ~   X 0.668 0.091 7.364  0.000    0.490    0.846     M ~ X
#> 32   Y  ~   M 0.307 0.080 3.829  0.000    0.150    0.464     Y ~ M
#> 33   M ~~   M 0.818 0.116 7.071  0.000    0.591    1.044    M ~~ M
#> 34   Y ~~   Y 0.811 0.115 7.071  0.000    0.586    1.035    Y ~~ Y
#> 35   X ~~   X 0.993 0.000    NA     NA    0.993    0.993    X ~~ X
#> 36   M  ~   X 0.346 0.110 3.141  0.002    0.130    0.561     M ~ X
#> 37   Y  ~   M 0.240 0.089 2.697  0.007    0.065    0.414     Y ~ M
#> 38   M ~~   M 1.051 0.149 7.071  0.000    0.760    1.342    M ~~ M
#> 39   Y ~~   Y 0.911 0.129 7.071  0.000    0.659    1.164    Y ~~ Y
#> 40   X ~~   X 0.868 0.000    NA     NA    0.868    0.868    X ~~ X
#> 41   M  ~   X 0.440 0.096 4.600  0.000    0.253    0.628     M ~ X
#> 42   Y  ~   M 0.280 0.101 2.775  0.006    0.082    0.477     Y ~ M
#> 43   M ~~   M 0.868 0.123 7.071  0.000    0.627    1.109    M ~~ M
#> 44   Y ~~   Y 1.069 0.151 7.071  0.000    0.773    1.365    Y ~~ Y
#> 45   X ~~   X 0.948 0.000    NA     NA    0.948    0.948    X ~~ X
#> 46   M  ~   X 0.286 0.104 2.741  0.006    0.081    0.490     M ~ X
#> 47   Y  ~   M 0.343 0.099 3.448  0.001    0.148    0.537     Y ~ M
#> 48   M ~~   M 1.131 0.160 7.071  0.000    0.817    1.444    M ~~ M
#> 49   Y ~~   Y 1.201 0.170 7.071  0.000    0.868    1.533    Y ~~ Y
#> 50   X ~~   X 1.041 0.000    NA     NA    1.041    1.041    X ~~ X